refactor(mcp): createFakeServer をテストヘルパーに共通化
packages/mcp/src/tools/meta-test-helpers.ts に createFakeServer,
registerDummyTool, ToolHandler, ToolResult を抽出し、
meta.spec.ts と meta.test.ts から参照する。

依存方向を正しく保つため、ヘルパーは packages/mcp 側に配置し、
spec からはパッケージエイリアス経由でインポートする。

Closes #577
